fix error when keys fail to resign
we had a bug where if the vault token became invalid (either due
to expiration or revokation) and we attempted to resign a key, we
would still try to update the key with the newly signed values
that were actually nil.

now if we fail to resign we leave the key untouched and attempt
again at the next cycle so long as the key is in the keychain.