validate the vault token when launching proxy
because of the ambiguity of the error message when a vault token
is invalid (expired, incorrect, etc) we should try to provide to
the user a more helpful error message if they attempt to spin up
a shell with an invalid token.

because we do not alert on errors once the shell has been launched
this only checks on launch and will not capture if the token
expires or is revoked after the shell is launched.